Adding functionality to transfer the win.
There's been quite a few instances where the OP hastily gave the win to the wrong person. This should allow the winner (who the OP replied to with +correct) to give the win away to anyone in that thread. 

As long as the next round hasn't been posted yet, it checks the most recent round for the winner replying "+give round" to any parent comment. If they do, it gives the author of that comment the win, then removes the win flair from the previous winner and gives it to the new winner. It also gives a fair play award to the person who gave up the round. (I don't know how to add it to the sidebar) It also changes the password, and gives it to the new winner. 

I have not bugtested the new code, so it would smart to look it over and test it before deploy.